tools/android/loading Re-enable autostart in the GCE startup script.
The test checking for the autostart parameter was assuming that GCE
would return the empty string when a metadata was not defined.
In fact, some error string is returned instead, and thus the value of
the string has to be explicitly checked.

Review URL: https://codereview.chromium.org/1844533002

Cr-Original-Commit-Position: refs/heads/master@{#384261}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76aa326bb5c62e334d2cc508f70bbde56532e56b