tools/android/loading Run main.py from an arbitrary directory
The path to analyze.py was defined as a relative path to the current
directory.
This CL changes this to get the path from the configuration instead.

The goal is to be able to launch the app from a directory that is not
under the Chromium checkout, and thus avoid polluting that checkout

Review URL: https://codereview.chromium.org/1848773002

Cr-Original-Commit-Position: refs/heads/master@{#384523}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b75ab05b606f449a858c2cc421994cf1fd2a228b