[Extensions] Remove the "use_movable_types" entry from idl/json files
All APIs are converted to using movable types, and it is the default in code
generation since r384649. We can remove the entry in the json/idl files.

Also print an error with the filename if a file fails to parse as json.

BUG=595949

Review URL: https://codereview.chromium.org/1848103005

Cr-Original-Commit-Position: refs/heads/master@{#384773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7b32419fe1b86cea29cad5dc1d54787f3d0ee88