[Extensions] Replace ExtensionRendererState with ExtensionApiFrameIdMap
Metrics* show that the ExtensionApiFrameIdMap has a 99.85% chance of
retrieving the proper data for a lookup on the IO thread (with no thread
hop). The ExtensionRendererState object only has a 77.67% chance. This
makes the map roughly 70x better. Add the window_id to the map info, and
remove ExtensionRendererState.

Note: Right now, we don't do a thread-hop to get the frame data if we
don't find it immediately, even though in theory this would give us
another 0.07% of frame. The reason is because adding a thread-hop here
is very negative for performance, and this is still ~70x better than our
previous.

*Metrics == Extensions.ExtensionFrameMapCacheHit,
            Extensions.ExtensionFrameMapLookupSuccessful,
            Extensions.ExtensionRendererStateCacheHit,
            Extensions.WebRequestEventFoundFrame

BUG=455776

Review URL: https://codereview.chromium.org/1847803004

Cr-Original-Commit-Position: refs/heads/master@{#385283}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1790e82aac659c28dfbd9aac82f1feb4a839f405