Record histogram for WKWebView's viewport rendering bug.
This CL records the number of times in which a page with no viewport
tag is rendered with an unusable zoom scale.  The metrics collected for
this histogram can be used to justify viewport tag injection if
necessary.

BUG=583231

Review URL: https://codereview.chromium.org/1759233002

Cr-Original-Commit-Position: refs/heads/master@{#385592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b8501536085bb3a00f26cad1c95a7873dc69891