Start implementing ForeignFetchEvent for foreign fetch events.
This implements most of the  new ForeignFetchEvent interface added in
https://github.com/slightlyoff/ServiceWorker/commit/d2fdb147c9c769ac6338117c1acd9f4010365f19

For now all this does is unwrap the Response from a ForeignFetchResponse
dictionary. In follow-up patches support for the other properties of a
ForeignFetchResponse will be added.

BUG=540509

Review URL: https://codereview.chromium.org/1845463002

Cr-Original-Commit-Position: refs/heads/master@{#385917}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3161394749bb5e100a58ac77aaafb7b9c240fcc3