Background task handler for grabbing UMA for background loading.
When we add the background task scheduler to wake up and check
conditions for background offline page loading, we want to get some UMA
to determine how well this will work for our users.

We measure if the device is on A/C power, how full the battery is, and
what type of networks the user has available.

BUG=

Review URL: https://codereview.chromium.org/1823623004

Cr-Original-Commit-Position: refs/heads/master@{#386268}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 30e113c3a0c7994aadb63801438dfed7a24443f2