[tools/perf] Make trybot bails out early if the benchmark is disabled on trybot platform
Note that this functionality only work to some certain
extent since the disabling can be very dynamic & requires
the actual hardware to know whether the benchmark is
actually disabled.

We bias on the side of not disabling the benchmark if we
don't have enough information to tell & add a warning
message telling users that the benchmark may be disabled.

BUG=591410

Review URL: https://codereview.chromium.org/1873623002

Cr-Original-Commit-Position: refs/heads/master@{#386410}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f209bc5e1f049ca2dad7ca1145a34332e3f27d7