Add histograms for observing UMA throttling effect.
Added a histogram for counting number of events when UMA log was not uploaded due to throttling and one more for counting number of dropped UMA logs.

Also, delete a histogram recording a connection type and actualy upload interval used to upload the previous UMA log which is not needed as we are done with the analysis of UMA logs on cellular.

BUG=599586

Review URL: https://codereview.chromium.org/1871733002

Cr-Original-Commit-Position: refs/heads/master@{#386440}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0417f978064cbeda802b99b21a88d126c66ff4da