Use is_valid_dump to check for valid dumps in telemetry exceptions.
Now that we have a property which stores whether or not a stack trace
is valid, we can use that to test for valid minidumps instead of looking
for platform dependent strings. This test should also work in Android
now so we can enable it on all platforms except for chromeos.

This CL also fixes an issue where asserts raised within an assertRaises
block would be eaten, so tests were not failing as expected.

R=nednguyen@google.com
TBR=jochen@chromium.org
BUG=561763

Review URL: https://codereview.chromium.org/1880333002

Cr-Original-Commit-Position: refs/heads/master@{#387386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6be24867adfaa4159909e2072778fbe3fc341fd