Clovis: unittesting fix and some general tweaks.
The root problem was that dependency_graph_unittest called
request_dependencies_lens.TestReqeusts.CreateLoadingTrace, and then changed some
of the timing of those requests. As those requests were global, that meant that
if a different unittest (eg, prefetch_view_unittest) happened to use the same
TestRequests, it would see the changed timing, and so run differently than if it
were run in isolation.

A nice way to fix that is to serialize and deserialize the trace, which exposed
some other holes in our organization, namely abstract classes that defined empty
methods instead of abstract methods (in the python world, that means using
"pass" instead of "raise NotImplementedError") and then some other gaps in our
serialization methods.

In order to diagnose & fix this, run_tests was extended to allow for multiple
tests to be specified. This does not change existing behavior when a single
argument is passed to run_tests, and does the right thing for multiple arguments
instead of silently ignoring them.

Review URL: https://codereview.chromium.org/1892073002

Cr-Original-Commit-Position: refs/heads/master@{#387590}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 046e2f3a95ac7bb7afd3b16bb18007c2bf71fd8e