clovis: Improve the handling of tracing categories.
The baseline categories exclude 'cc', drastically reducing the size of a
loading trace (85 -> 39MB in one sample case), and are suitable for all
tools but sandwich. This one adds an extra category which is too heavy
to be in the default set.

Review URL: https://codereview.chromium.org/1901653002

Cr-Original-Commit-Position: refs/heads/master@{#388446}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3bccc967be8d4d4e27f2f7449b9c63ad5fc440a