File API: Add a use counter for nonstandard 'endings' option
Chrome supports the nonstandard 'endings' option for the Blob and File
constructors, which normalizes CR in the data to CR+LF on Windows.
It'd be nice to eliminate this if we can, so let's measure the usage.

BUG=605710

Review URL: https://codereview.chromium.org/1906153002

Cr-Original-Commit-Position: refs/heads/master@{#389055}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec90317104f275c3e818f1a6e4d800ce5c8d6990