Add support for cros chrome_sdk (simplechrome) to MB.
When we build Chrome for ChromeOS using the simplechrome
workflow, all of the GYP_DEFINES and gn args that are needed
are determined by the cros wrapper scripts, and rather than
duplicate that logic in the MB mixins, it's easier to just
pass the data through.

However, we don't want this passthrough mechanism to be abused,
so we implement so that this only works if the bot is configured
to allow for passthrough *and* the GYP_DEFINES or GN_ARGS env
vars contain chromeos=1 or target_os="chromeos" as appropriate.

If we need to be less strict later we can alway change things.

R=stevenjb@chromium.org
BUG=605154

Review URL: https://codereview.chromium.org/1913703002

Cr-Original-Commit-Position: refs/heads/master@{#389532}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73ed0d6aa6f1c9c9a045090574d92493dc12307d