cc: Add TracingInfo to image decodes and a new histogram.
This patch adds TracingInfo to encapsulate any information that we pass
to image decode controllers for the sole purpose of tracing/recording
stats.

Additionall, this adds three new histograms:

Renderer4.LockExistingCachedImage.Software.NOW
Renderer4.LockExistingCachedImage.Software.SOON
Renderer4.LockExistingCachedImage.Software.EVENTUALLY

These histograms record whether or not we succeeded in locking an
existing scaled and cached image decode. The metric is split by bin
where the bin represents the tile priority bin which caused the image to
be requested in the first place. For example, if a SOON tile has an
image dependency that causes us to look up the image and try to lock it,
then this metric will be recorded in the
Renderer4.LockExistingCachedImage.Software.SOON histogram. Note that if
two or more tiles attempt to lock the same image during the same round
of scheduling, only the first one will attempt to lock it (this will
typically correspond to the more important bin). The reason for this is
that it will either succeed (then the next tiles will find a locked
image) or it will fail (then the cache entry will be removed and a new
decode started).

R=enne, ericrk
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1912243005

Cr-Original-Commit-Position: refs/heads/master@{#389625}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43bba8a1efe4e52e8948ebd0574eae3d1f946825