sandwich: Output further information in metrics' CSV
To make sure there is not misinterpretation of the metrics
from a sandwich CSV, this CL adds the network emulation column
in the CSVs.

This CL is also the opportunity to get ride of the sandwich's
run_info.json that is completely replaceable with LoadingTrace's
metadata dictionary.

BUG=582080

Review URL: https://codereview.chromium.org/1891183003

Cr-Original-Commit-Position: refs/heads/master@{#390078}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 929aa5f0811b7e9a3c66d8b59fb4ff2b95f0a243