tools/android/loading: Force to use of simple cache on desktop.
Before on desktop, the cache backend type of the HTTP cache was
choosen by the SimpleCacheTrial on field trial. This can cause
issues with sandwich were the cache backend type might change
between runs.

This CL override the cache backend type to use in Chrome on
Desktop to not have to handle the two backend types in the cache
archive processing.

BUG=582080

Review-Url: https://codereview.chromium.org/1931523002
Cr-Original-Commit-Position: refs/heads/master@{#390352}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e6c8e0d72463672af06ecd6a634e6d9e5078927