Enable dcheck on the GPU waterfall bots.
Since we can't afford to run the debug versions of the tests
on the GPU waterfall bots, we want to run release builds w/
dchecks enabled.

Note that we enable dchecks here, but the dcheck chromium config is
*not* applied on the build-side recipes. We do that intentionally
so that a given trybot can merge the configs from both the main
waterfall bots and the gpu bots (which otherwise would have conflicting
configs), but this may be confusing. The right fix here is to do
away with as many recipe-side configs that set GYP_DEFINES as possible,
since the real settings are governed by MB, src-side.

R=kbr@chromium.org, phajdan.jr@chromium.org
BUG=606700

Review-Url: https://codereview.chromium.org/1933533002
Cr-Original-Commit-Position: refs/heads/master@{#390828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe42ebe1aa800787216d69be07b52411076592dd