[battor agent] Makes the main() thread also be the UI thread
This requires changes to the threading model because, before, we were
allowed to block the main() thread until each of the tracing commands
returned. Now that we're using the main() thread as the UI thread,
there's an expectation that it won't block while the command is
executing (because at some point the IO thread might delegate back to
the IO thread for a task, resulting in a deadlock).

This fixes a bug on Windows where a MessageWindow that was created on
the UI thread was check failing on exit because it was expecting to be
destroyed on the same thread. This happened every time that the agent
shut down on Windows.

BUG=603309

Review-Url: https://codereview.chromium.org/1949673003
Cr-Original-Commit-Position: refs/heads/master@{#391599}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0d672480e8c352a0f7ab4d67a87d58bb543d943