Desktop Capture Picker New UI: Non Mac Structure Change
A re-commit of 1932413002.
The reason of unittest crash is that we removed a placeholder
in the string resource for window title, but we still used
GetStringFUTF16 in the code of old UI.

In this CL, we add a deprecated string resource with placeholder
for old picker UI to use.

BUG=602478, 610016
TBR=isherman@chromium.org,rockot@chromium.org

Review-Url: https://codereview.chromium.org/1966653002
Cr-Original-Commit-Position: refs/heads/master@{#392770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ced272025f70119b5747514bdb112591f36557e