DNS: Retain stale entries in HostCache and return when requested
In preparation for future connection optimizations, retain stale
entries (those with a response past the TTL or received on a different
network) in the cache, and expose a way to look up and characterize the
stale results.

To do this, HostCache moves away from using ExpiringCache to a self-
managed map of keys to entries, with extra metadata.

Also, add a bunch of histograms measuring how stale updated entries are.

BUG=605138

Review-Url: https://codereview.chromium.org/1908543002
Cr-Original-Commit-Position: refs/heads/master@{#393246}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 317860ed4f3e3788fd3c4637ee5d91afd31e39b1