tools/android/loading Add a task handler for generating trace reports
The report task:
- searches files with prefix "trace_database" and loads them.
- for each URL, lookup in the database and find the traces
- load the trace, generate a report from it
- upload to BigQuery using the streaming API. This creates
  the table lazily, but requires a table template (not
  created automatically).

The change of the URL in TraceReportHandler is to ensure
that the URL in the loading trace database is the same as
the one given by the user.
Otherwise, the user will not be able to find some URLs in
the database when generating reports.

Review-Url: https://codereview.chromium.org/1947123002
Cr-Original-Commit-Position: refs/heads/master@{#393500}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d47c255f3700621eeb9b8b5ba3897cb106b8364c