tools/android/loading Share path constants between frontend and backend
The frontend and the backend need to share knowledge about
file and URL paths (such as the path where the loading
trace databases are generated, and the BigQuery table URL).

This CL removes the duplication and moves the code to a
common location.

Review-Url: https://codereview.chromium.org/1979623002
Cr-Original-Commit-Position: refs/heads/master@{#394092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7305d984f9092c165c4b3ec9debb13336d51572c