[TTS] Add user actions for the Contextual Search panel.
Adds user actions for manipulation of the Contextual Search panel.
Actions include both what was done and how it was done, e.g. BasePageScrollClose.
Also logs a user action for viewing the contents of the panel that describes how long the contents were viewed.
Fixed an issue where a close was due to promotion but the internal reason was unknown.
All new actions are prefixed with "ContextualSearch." for scoping.

BUG=610820

Review-Url: https://codereview.chromium.org/1986673003
Cr-Original-Commit-Position: refs/heads/master@{#394591}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d0fe195c0a376db680823555c6be7e7afd68587