Sandwich: Skip SpeedIndex from CSV if cannot calculate
Sometimes the bounding box is not available on a video, I guess this is because
screenrecord misses a few frames. This happens rarely, so we'd mark these cases
as "unavailable" in the CSV and will watch for these not to happen too often to
skew results.

BUG=582080

Review-Url: https://codereview.chromium.org/2007593002
Cr-Original-Commit-Position: refs/heads/master@{#395372}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dabb3564f823df7b0aae8faf220bee2a59649a3