[PCv1] Add timeToOnload metric for PCv2 compat testing.
This CL adds "timeToOnload" metric, which measures time from
navigationStart to onload event.
This metric is also added to PCv2 in CL:
https://codereview.chromium.org/2005753002/

This is a temporary metric to measure PCv1/PCv2 compatibility.
This metric is to be removed after the compatibility testing is complete.

BUG=catapult:#2357

Review-Url: https://codereview.chromium.org/1997383002
Cr-Original-Commit-Position: refs/heads/master@{#395879}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be36b308eab9cd9042ded805838267b0c53a929d