Implementation of new policy, that will specify the domains of user accounts allowed to log in to GAIA from a managed device.
Policy name is AllowedDomainsForApps. It is defined as a string, comma separated list of domains that are allowed.
When empty or not set, all the accounts are allowed to be used.

When policy is present, browser includes additional header with name X-GoogApps-Allowed-Domains and value from the policy in every request that user does.
If request is directed to GAIA service, it will be checked against the URL to see if the domain is allowed.

This can be tested, setting the policy manually on local machine (for example on linux a file /etc/chromium/policies/managed/allowedDomainsPolicy.json
with contents { "AllowedDomainsForApps": "google.com" }, running the chromium browser and trying to login in gmail with some gmail account.

The page received will say that the admin doesn't allow to use that account on this machine.

BUG=470521

Review-Url: https://codereview.chromium.org/1659623004
Cr-Original-Commit-Position: refs/heads/master@{#396436}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 193d8ad34333979c0551d22e06800d8f55a5edde