tools/android/loading: Consolidate code tracking per-user-lens metrics
This CL refactors report.py in order to abstract the notion of tracking
a bunch of metrics for a variety of different user lenses (e.g., first
text/first contentful paint/first significant paint). It introduces
a new PerUserLensReport class that takes in a user lens instance and
tracks a variety of data for that user lens. LoadingReport is changed to
use one instance of this class per desired user lens rather than
computing each individual metric manually per user lens.

The intent is to simplify the addition of either future user lenses or
future per-user-lens metrics: Each of those additions can now occur at
just one place in the code.

This change is not backwards compatible, as it changes the names of some
of the data generated in a loading report.

Review-Url: https://codereview.chromium.org/2016333002
Cr-Original-Commit-Position: refs/heads/master@{#396454}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0066d5542c87bc9d7a0fbefbed54a19a4f0ebdf