Update Maps pixel test.
Pick up a new binary from the Maps team, and update the test
expectations for a differently-sized viewport.

It looks like the sizing of the canvas in this test is slightly
different on Android than on desktop platforms, despite the fact that a
fixed device pixel ratio is sent to the Maps test itself. Work around
this by adding a mechanism to cloud_storage_test_base's test
expectations for device-specific scale factors. The machine names are
already being passed down to the test on the bots, so this will allow
the same expectations to apply on all devices. Tested locally on a Nexus
6 and 9.

Removed the failure expectations for this test (which has now been
renamed to Maps.maps_004) on Android. The test will probably fail on the
Nexus 5 and 5X, but scale factors will be checked in for these devices
in a follow-on CL.

BUG=611932
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_optional_gpu_tests_rel;tryserver.chromium.mac:mac_optional_gpu_tests_rel;tryserver.chromium.win:win_optional_gpu_tests_rel

Review-Url: https://codereview.chromium.org/2010363002
Cr-Original-Commit-Position: refs/heads/master@{#396541}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c329cfaedf7a947ea84e31e90c9112f4d64c3e24