tools/android/loading Allow changing the output directory
Task output is now generated in a unique directory, avoiding
overriding the results from previous runs.

The name of the unique directory is composed of the name of the
user, a unique ID, and optionally a task name.

The 'report' task now needs an extra parameter, since the trace
directory is no longer hardcoded.

Review-Url: https://codereview.chromium.org/2008033003
Cr-Original-Commit-Position: refs/heads/master@{#396692}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06e4163835bf23a8d5eb7d9ddce33024dec98316