[sensors]: Introduce the Generic Sensor API.
The Generic Sensor API defines a framework for exposing sensor data to
the Open Web Platform. This CL introduces the IDLs for Sensor* interface
and adds the supporting .h|cpp files.

The ED for the Generic Sensor API is here -
https://w3c.github.io/sensors/

Intent to Implement :
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/TkfdVqYAYiE

BUG=606766

Review-Url: https://codereview.chromium.org/1942663003
Cr-Original-Commit-Position: refs/heads/master@{#396707}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea88054ae4e7cd154177bb23426271fc0d6906ab