[HttpCache] Record age of validated and updated stale cache entries
It would be good to know the correlation between an entry's age and its
likelihood to be validated vs updated. This CL adds two new metrics which
record the stale entry's age (as a percentage of freshness) when it's validated
and when it's updated.

BUG=614355

Review-Url: https://codereview.chromium.org/2036123003
Cr-Original-Commit-Position: refs/heads/master@{#397720}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc3cb4f7c2ca1f0ade7dcb560a8040ed2d13a695